+++ lib/puppeteer_wrapper/puppeteer_page.ts
@@ -152,4 +152,8 @@ export default class PuppeteerPage {
     public setExtraHTTPHeaders(headers: Record<string, string>): Promise<void> {
         return this.page.setExtraHTTPHeaders(headers);
     }
+
+    public setContent(html: string): Promise<void> {
+        return this.page.setContent(html);
+    }
 }