+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+2.7.0 / 2019-10-20
+==================
+
+* Browser.setContent
+* LogRequests option
+* Updated types of dependency
+* Removed light versions of logo
+
 2.6.3 / 2019-10-10
 ==================